pyproject.toml
[tool.poetry]
version = "0.1.0"
requires-poetry = '>=2.0'

packages = [
    { include = "src" },
]

[tool.poetry.scripts]
app = "src.main:main"

[tool.poetry.dependencies]
requests = "^2.31.0"
ujson = "^5.8.0"
pynacl = "1.5.0"
yt-dlp = "2026.2.21"
async-lru = "^2.0.4"
async-timeout = "^5.0.1"
discord-py = "2.6.4"
black = "25.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore:'audioop' is deprecated:DeprecationWarning",
]

[project]
name = "discord-music-bot"
description = ""
dynamic = ["version", "dependencies"]
readme = "README.md"
authors = []
maintainers = []
requires-python = '>=3.11,<4.0'
